cargo_zng::fmt

Function custom_fmt_rs

async fn custom_fmt_rs(
    rs_file: PathBuf,
    check: bool,
    fmt: FmtFragServer,
) -> Result<Option<PathBuf>>

Applies custom format for all macro bodies in file, if changed writes the file and returns the file path for reformat.

Changed files need reformat in cases like [Macro!{..}, Macro!{..}] where the custom macro format introduces line break, this causes rustfmt to also make the [] multiline.