cargo_zng::fmt

Module pm2_send

proc_macro2 types are not send, even when compiled outside of a proc-macro crate this mod converts the token tree to a minimal Send model that only retains the info needed to implement the custom formatting

Structs

Group

Ident

Punct

Span

TokenStream

Enums

Delimiter

Describes how a sequence of token trees is delimited.

TokenTree