cargo_zng::l10n

Module scraper

Localization text scraping.

Structs

CommandMacroArgs 🔒

CommandMacroEntry 🔒

CommandMetaEntry 🔒

FluentEntry

Represents one call to l10n! or similar macro in a Rust code file.

FluentNote

Represents a standalone note, declared using // l10n-{file}-### {note} or l10n-### {note}.

FluentTemplate

Represents all calls to l10n! or similar macro scraped from selected Rust code files.

L10nMacroArgs 🔒

Functions

parse_validate_id 🔒

scrape_file 🔒

scrape_files 🔒

scrape_fluent_text

Scrapes all use of the l10n! macro in Rust files selected by a glob pattern.