cargo_zng::l10n::scraper

Function scrape_fluent_text

pub fn scrape_fluent_text(
    code_files_glob: &str,
    custom_macro_names: &[&str],
) -> FluentTemplate

Scrapes all use of the l10n! macro in Rust files selected by a glob pattern.

The custom_macro_names can contain extra macro names to search in the form of the name literal only (no :: or !).

Scraper does not match text inside doc comments or normal comments, but it may match text in code files that are not linked in the Cargo.toml.

See FluentEntry for details on what is scraped.

Panics

Panics if code_files_glob had an incorrect pattern.