Module cargo_zng::util

Macros

error 🔒

Print error message and flags the current process as failed.

fatal 🔒

Print error message and exit the current process with error code.

warn 🔒

Print warning message.

Statics

ERROR_PREFIX

RUN_FAILED 🔒

WARN_PREFIX

Functions

cmd

Run the command with args.

exit

Exit the current process, with error code 102 if is_failed_run.

is_failed_run

Gets if the current process will exit with error code.

set_failed_run

Sets if the current process will exit with error code.

workspace_dir