Module cargo_zng::util

Macros

error 🔒

Print error message and flags the current process as failed.

fatal 🔒

Print error message and exit the current process with error code.

warn 🔒

Print warning message.

Structs

DependencyManifest

Statics

ERROR_PREFIX

RUN_FAILED 🔒

WARN_PREFIX

Functions

ansi_enabled

check_or_copy

check_or_create_dir

check_or_create_dir_all

check_or_write

clean_value

cmd

Run the command with args, inherits stdout and stderr.

cmd_impl 🔒

cmd_silent

Run the command with args.

dependencies

Workspace root and dependencies of manifest_path

exit

Exit the current process, with error code 102 if is_failed_run.

is_failed_run

Gets if the current process will exit with error code.

manifest_path_from_package

set_failed_run

Sets if the current process will exit with error code.

workspace_dir

workspace_manifest_paths

Workspace crates Cargo.toml paths.